test_kgdb_selftest: Tidy up xfail and timeout logic
Added an xfail for ARCH=arm. I went back and forth over the kernel
versions and cannot find one where this test passes. Not clear why
it didn't have an xfail already.

We also clean up the timeout logic to allow us to use a relatively short
timeout on >v4.17 kernels (which include code to issue progress reports
while the self test is running). This allows us to keep the test
runnable despite the xfail.

Signed-off-by: Daniel Thompson <daniel@redfelineninja.org.uk>