kbuild/ktest: Allow gdb to be overridden by the environment
When a new toolset introduces a regression it is often useful (for
blame allocation) to compile with one toolset but use gdb from the other.
These changes makes is possible to use CROSS_GDB to force the test
suite to use a special version of gdb without changing the impact of
CROSS_COMPILE on kbuild.

Signed-off-by: Daniel Thompson <daniel@redfelineninja.org.uk>