kgdb_boottest: Improve test coverage (and update xfail)
There's a degree of hackery here but currently all kgdbts runs fail for
the same reason on x86 (because hw_break doesn't trap proplerly on data
access) and that means we end up ignoring the latest test coverage...

Modify the error handling to capture the reason and ignore hw_break
errors (which will still be reported by kgdb_selftest/kgdbts_V1). Sadly
this reveals a further problem with stepping through the boot functions
which, for now, we handle by describing it with an xfail.

Signed-off-by: Daniel Thompson <daniel@redfelineninja.org.uk>