test_kgdb_boottest: Don't expect to enter kdb
Currently the test_kgdbts_boot() explicitly expects an Entering kdb message
to be issued as kgdbts starts to run. This message is only issued in a
user visible manner if there is a console handler registered and we are
starting to see Arm systems register their console handlers much later in
the boot process than this historical norm (nowadays we can rely on
earlycon to debug boot problems so a late console isn't as bad as it used
to be).

Signed-off-by: Daniel Thompson <daniel.thompson@linaro.org>