ktest: Improve sub-process management
When running on systems with only a single serial port (e.g. arm64/riscv
virt  machines) we can get some really nasty races waiting for picocom
to be fully ready. If we don't wait for picocom to be ready before
allowing qemu to proceed then we lose boot messages (and expext_boot()
will fail).

I haven't yet found a robust workaround that does not rely on sleeping :-(.
Even reaching the initial select isn't enough for the serial ports to be
fully configured! Eeek!

Signed-off-by: Daniel Thompson <daniel.thompson@linaro.org>