tests: Fix everything that uses kill
Currently tests that run processes in the background and then clean them
up using kill are unreliable (about 1 in 30 failure rate on my current
build server). It appears this is because the prompt can be concealed by
messages telling us the tasks have terminated (although it isn't entirely
clear why... most shells do not asynchronously report termination to avoid
exactly this problem).

So it goes. We can avoid this problem by changing our expectations a little.

Signed-off-by: Daniel Thompson <daniel.thompson@linaro.org>