ktest: Ensure run_command_kdb reliably eats the initial line ending
Currently, when we run a command, we expect to see the command echoed back
to us followed by [\r\n]*. This pattern makes it unpredictable whether or
not there will be a leading \n in the output (the glob will match if the \n
has been read but not if the DUT has not yet issued it).

Fix this by removing the repeat in the regular expression and striping any
leading '\n' before handing back the output.

Signed-off-by: Daniel Thompson <daniel.thompson@linaro.org>